edited .gitignore
