+++ .gitignore
@@ -8,6 +8,7 @@
 # Test Suite
 /test/testdocs/
 /test/env/
+/test/logs/
 
 # Cache
 *__pycache__*
